Fix setting odata.type via constructor
We only want to set odata.type when 1. a type is a derived type, and
2. one of its base types is referenced as the type for a property in
another type. The odata.type is necessary for the derivedtypeconverter
to properly deserialize the type. Current behavior is that we set
odata.type on all entities and complex types. Some service APIs do not
properly handle the unexpected instance annotation, even though they
should ignore it.

On branch bugfix/odataTypeCstor
Changes to be committed:
	modified:   Templates/CSharp/Model/ComplexType.cs.tt
	modified:   Templates/CSharp/Model/EntityType.cs.tt
	modified:   src/GraphODataTemplateWriter/Extensions/OdcmModelExtensions.cs